requirements.txt
click==7.1.2
PyQt5==5.15.1
pyqt5-plugins==5.15.1.2.0.1
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
pyqt5-tools==5.15.1.3
PyQt5Designer==5.14.1
pyserial==3.5
python-dotenv==0.15.0
qt5-applications==5.15.1.2
qt5-tools==5.15.1.1.0.1
